Adicionei checagem de coluna a migração
Creio que tenha resolvido erro que ocorria toda vez que as migrações
eram rodadas. O erro era a tentativa de criar a coluna course_id em
materiais, sendo que ela já existia. Agora há uma checagem que evita
isso.

+++ database/migrations/2020_04_29_005405_remove_course_fk_from_objects.php
@@ -13,23 +13,30 @@ class RemoveCourseFkFromObjects extends Migration
      */
     public function up()
     {
-      Schema::table('learning_objects', function (Blueprint $table) {
-        // $table->dropForeign('courses_learning_objects_id_foreign'); <- fallback
-        // $table->dropForeign(['course_id']); maybe we don't need to drop it, dunno.
-        $table->dropColumn('course_id');
-      });
+      $columnExists = Schema::hasColumn('learning_objects', 'course_id');
+
+      if ($columnExists) {
+        Schema::table('learning_objects', function (Blueprint $table) {
+          $table->dropColumn('course_id');
+        });
+      }
     }
 
-    /**R
+    /**
      * Reverse the migrations.
      *
      * @return void
      */
     public function down()
     {
-      Schema::table('learning_objects', function(Blueprint $table) {
-        $table->integer('course_id')->unsigned();
-        $table->foreign('course_id')->reference('id')->on('courses');
-      });
+      $columnExists = Schema::hasColumn('learning_objects', 'course_id');
+
+      if (!$columnExists) {
+        Schema::table('learning_objects', function(Blueprint $table) {
+          $table->integer('course_id')->unsigned();
+          $table->foreign('course_id')->reference('id')->on('courses');
+        });
+      }
     }
 }
+
